refactor/ apply mvc pattern
