--- src/main/java/calculator/Calculator.java
@@ -1,11 +1,0 @@
-package calculator;
-
-public class Calculator {
-  public static int calculate(int[] numbers) {
-    int sum = 0;
-    for (int number : numbers) {
-      sum += number;
-    }
-    return sum;
-  }
-}
