--- src/main/java/calculator/Controller.java
@@ -1,2 +1,0 @@
-package calculator;public class Controller {
-}
