--- src/main/java/calculator/OutputController.java
@@ -1,9 +1,0 @@
-package calculator;
-
-import calculator.util.Constant;
-
-public class OutputController {
-  public static void printResult(int result){
-    System.out.println(Constant.OUTPUT_MESSAGE + result);
-  }
-}
