+++ src/main/java/calculator/dto/ParsedResult.java
